Moved canvas to OOP, major rework, added straight line and rectangle tools
