+++ js/classes/Point.js
@@ -1,0 +1,6 @@
+export default class Point {
+    constructor(x, y) {
+        this.x = x;
+        this.y = y;
+    }
+}
