+++ js/uiHelpers.js
@@ -1,10 +1,1 @@
-document.querySelectorAll("[data-tool").forEach((tool) => {
-  tool.addEventListener("click", (e) => {
-    toggleClassOnTool(e.target, "selected");
-  });
-});
 
-function toggleClassOnTool(tool, className) {
-  document.querySelector("[data-tool].selected").classList.remove(className);
-  tool.classList.toggle(className);
-}
